+++ src/main/java/lotto/service/LottoPurchaseService.java
@@ -1,7 +1,7 @@
 package lotto.service;
 
-import lotto.domain.lotto.Lotto;
-import lotto.domain.lotto.Lottos;
+import lotto.domain.Lotto;
+import lotto.domain.Lottos;
 import lotto.exception.LottoException;
 import lotto.exception.LottoStoreException;
