+++ src/main/java/lotto/service/LottoWinningRankingService.java
@@ -1,9 +1,9 @@
 package lotto.service;
 
-import lotto.domain.lotto.Lotto;
-import lotto.domain.lotto.Lottos;
-import lotto.domain.winning.LottoWinningRanking;
-import lotto.domain.winning.LottoWinningSet;
+import lotto.domain.Lotto;
+import lotto.domain.Lottos;
+import lotto.domain.LottoWinningRanking;
+import lotto.domain.LottoWinningSet;
 import lotto.vo.LottoNumber;
 
 import java.util.EnumMap;
